+++ rollup.test.config.js
@@ -3,7 +3,7 @@ import { warningHandler as onwarn } from '@shgysk8zer0/js-utils/rollup';
 import { rollupImport } from '@shgysk8zer0/rollup-import/import';
 import terser from '@rollup/plugin-terser';
 import { rollupImportMeta } from '@shgysk8zer0/rollup-import/meta';
-import { importmap } from './importmap.js';
+import { importmap } from '@shgysk8zer0/importmap';
 import { readJSONFile } from '@shgysk8zer0/npm-utils/json';
 
 const { name, version } = await readJSONFile('./package.json');